added notify-logged
