+++ packages/ui-contexts/src/index.ts
@@ -85,14 +85,8 @@ export { useIsDeviceManagementEnabled } from './hooks/useIsDeviceManagementEnabl
 export { useSetOutputMediaDevice } from './hooks/useSetOutputMediaDevice';
 export { useSetInputMediaDevice } from './hooks/useSetInputMediaDevice';
 
-export {
-	StreamerEvents,
-	ServerMethods,
-	ServerMethodName,
-	ServerMethodParameters,
-	ServerMethodReturn,
-	ServerMethodFunction,
-} from './ServerContext/methods';
+export { ServerMethods, ServerMethodName, ServerMethodParameters, ServerMethodReturn, ServerMethodFunction } from './ServerContext/methods';
+export { StreamerEvents } from './ServerContext/streams';
 export { UploadResult } from './ServerContext';
 export { TranslationKey, TranslationLanguage } from './TranslationContext';
 export { Fields } from './UserContext';
